+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Fixed
 * Only set ``auto_refresh_url`` in ``make_google_blueprint`` if a token of
   type ``offline`` is requested. See issues `#143`_, `#144`_ and `#161`_ for
   background.
+* Added ``hd`` argument to ``make_google_blueprint``
 
 `1.0.0`_ (2018-06-04)
 ------------------